+++ RedPandaIDE/translations/RedPandaIDE_pt_BR.ts
@@ -1402,11 +1402,31 @@
         <translation>Usar fonte personalizada</translation>
     </message>
     <message>
-        <source>Show special chars</source>
+        <source>Enable ligatures support</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Enable ligatures support</source>
+        <source>Line Spacing:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Show whitespaces</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Leading</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Inner</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Trailing</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Line break</source>
         <translation type="unfinished"></translation>
     </message>
 </context>